[EI-261] [api] feat: 가중치 적용 VAR 분석 기능 추가
