+++ apps/data-modeling-api/main.py
@@ -1,7 +1,6 @@
 from fastapi import Depends, HTTPException, Query
 from fastapi import FastAPI
-# from service import predict, sourceIndicatorsVerification, predictWithoutSourceIndicators
-from service2 import sourceIndicatorsVerification, predict
+from service import sourceIndicatorsVerification, predict
 from database import engine, Base, get_db
 from sqlalchemy.orm import Session
 from mangum import Mangum
@@ -39,11 +38,10 @@ def loadSourceIndicatorsVerification(
     targetIndicatorType: str = Query(...),
     sourceIndicatorId: list[str] = Query(default = None),
     sourceIndicatorType: list[str] = Query(default = None),
-    weight: list[int] = Query(default = None),
     db: Session = Depends(get_db)
     ):
       try:
-        verificaion = sourceIndicatorsVerification(targetIndicatorId, targetIndicatorType, sourceIndicatorId, sourceIndicatorType, weight, db)
+        verificaion = sourceIndicatorsVerification(targetIndicatorId, targetIndicatorType, sourceIndicatorId, sourceIndicatorType, db)
         return verificaion
       except Exception as error:
         raise HTTPException(status_code=500, detail=f"{str(error)}")
